simplify document module by moving cruft and miscellanea to document.util
