+++ bokeh/document/__init__.py
@@ -2,6 +2,7 @@
 
 
 '''
+from __future__ import absolute_import
 
 from .document import Document ; Document
