+++ bokeh/document/locking.py
@@ -1,6 +1,7 @@
 '''
 
 '''
+from __future__ import absolute_import
 
 from ..util.future import wraps
